Separate MessageArea component
